Use public queries and move tests
