+++ contracts/cw1155-base/src/lib.rs
@@ -4,3 +4,6 @@ pub mod msg;
 pub mod state;
 
 pub use crate::error::ContractError;
+
+#[cfg(test)]
+mod tests;
